Former-commit-id: 8f7a1ea4182163bee835af5317f143276ec368d4

+++ backend/Foody/src/main/java/com/foody/member/dto/request/MemberJoinRequest.java
@@ -2,8 +2,8 @@ package com.foody.member.dto.request;
 
 public record MemberJoinRequest(
     String nickname,
-    float height,
-    float weight,
+    double height,
+    double weight,
     int gender,
     int age,
     int activityLevel
